Update README.md
This program put date, wage, wage per hour and sum of money earned for a day and the total earned together to create a file describe them. I was not able to make the total in one go because I don't know how to get the sum before sum shown up, so I created a separate file to contain that information to make the logic workable.